Revert "prime-factors: replace u32 with u64 in stub"
This reverts commit 3c8b9bb5bd95c0aa5931d1e6ab343ee3b2b400ea.

+++ exercises/prime-factors/src/lib.rs
@@ -1,3 +1,3 @@
-pub fn factors(n: u64) -> Vec<u64> {
+pub fn factors(n: u32) -> Vec<u32> {
     unimplemented!("This should calculate the prime factors of {}", n)
 }
